Rename galleria view to slideshow
