+++ src/collective/nitf/locales/collective.nitf.pot
@@ -77,7 +77,7 @@ msgstr ""
 msgid "Body text"
 msgstr ""
 
-#: collective/nitf/templates/galleria.pt:42
+#: collective/nitf/templates/slideshow.pt:42
 msgid "By"
 msgstr ""
 
@@ -106,7 +106,7 @@ msgid "Default Urgency"
 msgstr ""
 
 #: configure.zcml
-msgid "Displays a gallery in the news article"
+msgid "Displays a slideshow in the news article"
 msgstr ""
 
 #: configure.zcml
@@ -114,7 +114,7 @@ msgid "Displays news article with image"
 msgstr ""
 
 #: collective/nitf/templates/media.pt:111
-msgid "Drag and drop images to change their order on the gallery."
+msgid "Drag and drop images to change their order on the slideshow."
 msgstr ""
 
 #: collective/nitf/config.py:52
@@ -142,7 +142,7 @@ msgid "From the Scene"
 msgstr ""
 
 #: configure.zcml
-msgid "Gallery View"
+msgid "Slideshow View"
 msgstr ""
 
 #: collective/nitf/config.py:194
