+++ src/collective/nitf/profiles/default/types/collective.nitf.content.xml
@@ -40,7 +40,7 @@
  <property name="default_view_fallback">False</property>
  <property name="view_methods">
   <element value="view"/>
-  <element value="galleria"/>
+  <element value="slideshow"/>
  </property>
 
  <!-- Method aliases -->
