+++ src/collective/nitf/static/nitf.js
@@ -4,17 +4,17 @@
     // If this is a NITF content with an image
     var nitf_view = $('body.portaltype-collective-nitf-content.template-view').length > 0;
     var nitf_with_image = nitf_view && $('.newsImageContainer').length > 0;
-    var nitf_galleria = $('body.portaltype-collective-nitf-content.template-galleria').length > 0;
+    var slideshow = $('body.portaltype-collective-nitf-content.template-slideshow').length > 0;
     if (nitf_with_image) {
       // Turn image link into something special
       var link = $('#parent-fieldname-image');
       // Add a magnify icon over the image
       link.append('<span class="magnify"></span>');
-      // For mobile devices, link will open galleria
+      // For mobile devices, link will open slideshow
       if (/Android|webOS|iPhone|iPad|iPod|BlackBerry/i.test(navigator.userAgent)) {
-        link.attr('href', link.attr('href').replace('@@galleria', '@@galleria?ajax_include_head=1&amp;ajax_load=1'));
+        link.attr('href', link.attr('href').replace('@@slideshow', '@@slideshow?ajax_include_head=1&amp;ajax_load=1'));
       } else {
-        // For other devices open a galleria in an overlay
+        // For other devices open a slideshow in an overlay
         link.prepOverlay({
           subtype: 'ajax',
           filter: '#content > *',
@@ -22,13 +22,13 @@
           config: {
             onLoad: function (e) {
               // Start cycle2
-              $('#galleria').cycle();
+              $('#slideshow').cycle();
             }
           }
         });
       }
-    } else if (nitf_galleria) {
-      // For galleria template, hide address bar after page load
+    } else if (slideshow) {
+      // For slideshow template, hide address bar after page load
       // in mobile devices
       /* To hide address bar after page load */
       var hideAddressBar = function() {
