+++ src/collective/nitf/tests/test_upgrades.py
@@ -165,7 +165,7 @@ class to2000TestCase(UpgradeTestCaseBase):
 
         # simulate state on previous version
         view_methods = list(fti.view_methods)
-        view_methods.remove('galleria')
+        view_methods.remove('slideshow')
         view_methods.append('nitf_galleria')
         fti.view_methods = tuple(view_methods)
         self.assertEqual(len(fti.view_methods), 2)
@@ -175,7 +175,7 @@ class to2000TestCase(UpgradeTestCaseBase):
         self._do_upgrade_step(step)
 
         self.assertEqual(len(fti.view_methods), 2)
-        self.assertItemsEqual(fti.view_methods, ['view', 'galleria'])
+        self.assertItemsEqual(fti.view_methods, ['view', 'slideshow'])
 
     def test_update_news_articles_layouts(self):
         title = u'Update News Articles layouts'
@@ -193,7 +193,7 @@ class to2000TestCase(UpgradeTestCaseBase):
         # execute upgrade step and verify changes were applied
         self._do_upgrade_step(step)
 
-        self.assertEqual(n1.getLayout(), 'galleria')
+        self.assertEqual(n1.getLayout(), 'slideshow')
 
     def test_install_new_dependencies(self):
         # check if the upgrade step is registered
