+++ src/collective/nitf/upgrades/v2000/profile/types/collective.nitf.content.xml
@@ -5,6 +5,6 @@
   </property>
   <property name="view_methods" purge="False">
     <element value="nitf_galleria" remove="True" />
-    <element value="galleria" />
+    <element value="slideshow" />
   </property>
 </object>
